Made unit tests pass for new service calls.  Moved test classes to own files.  Still contains issue where some fail when run all at once.
