+++ MessagingServiceTests/MessagingServiceTests.csproj
@@ -116,9 +116,16 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
+    <Compile Include="ChangeEmailNoticeTests.cs" />
+    <Compile Include="ContactOwnersTests.cs" />
+    <Compile Include="ContactSupportTests.cs" />
+    <Compile Include="EditCredentialTests.cs" />
+    <Compile Include="InvitePackageOwnerTests.cs" />
     <Compile Include="MessagingServiceGETTests.cs" />
-    <Compile Include="MessagingServicePOSTTests.cs" />
+    <Compile Include="NewAccountWelcomeTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ReportAbuseTests.cs" />
+    <Compile Include="ResetPasswordInstructionsTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\src\NuGet.Services.Messaging\NuGet.Services.Messaging.csproj">
